Restructure table files away from repository
