+++ manager/poll_service.py
@@ -4,7 +4,7 @@ from discord import Colour, Embed, Guild
 from discord.ext import commands
 
 from repository.poll_repo import PollRepository
-from repository.poll_table import (PollAnswerModel, PollMemberAnswerModel,
+from repository.table.poll_table import (PollAnswerModel, PollMemberAnswerModel,
                                    PollModel, VoteType)
