+++ repository/poll_repo.py
@@ -1,7 +1,7 @@
 from sqlalchemy import delete, join, select
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
-from repository.poll_table import (PollAnswerModel, PollGuildModel,
+from repository.table.poll_table import (PollAnswerModel, PollGuildModel,
                                    PollMemberAnswerModel, PollModel, VoteType)
