+++ repository/table/game_lobby_tables.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from sqlalchemy import Boolean, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from repository.db_config import Base
+from ..db_config import Base
 
 
 class GuildModel(Base):
